src/main/webapp/app/entities/comentario/comentario.service.ts
import { Injectable } from '@angular/core';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import * as moment from 'moment';
// eslint-disable-next-line @typescript-eslint/no-unused-vars
import { DATE_FORMAT } from 'app/shared/constants/input.constants';
import { map } from 'rxjs/operators';

import { SERVER_API_URL } from 'app/app.constants';
import { createRequestOption } from 'app/shared/util/request-util';
import { IComentario } from 'app/shared/model/comentario.model';

type EntityResponseType = HttpResponse<IComentario>;
type EntityArrayResponseType = HttpResponse<IComentario[]>;

@Injectable({ providedIn: 'root' })
export class ComentarioService {
  public resourceUrl = SERVER_API_URL + 'api/comentarios';

  constructor(protected http: HttpClient) {}

  create(comentario: IComentario): Observable<EntityResponseType> {
    const copy = this.convertDateFromClient(comentario);
    return this.http
      .post<IComentario>(this.resourceUrl, copy, { observe: 'response' })
      .pipe(map((res: EntityResponseType) => this.convertDateFromServer(res)));
  }

  update(comentario: IComentario): Observable<EntityResponseType> {
    const copy = this.convertDateFromClient(comentario);
    return this.http
      .put<IComentario>(this.resourceUrl, copy, { observe: 'response' })
      .pipe(map((res: EntityResponseType) => this.convertDateFromServer(res)));
  }

  find(id: number): Observable<EntityResponseType> {
    return this.http
      .get<IComentario>(`${this.resourceUrl}/${id}`, { observe: 'response' })
      .pipe(map((res: EntityResponseType) => this.convertDateFromServer(res)));
  }

  query(req?: any): Observable<EntityArrayResponseType> {
    const options = createRequestOption(req);
    return this.http
      .get<IComentario[]>(this.resourceUrl, { params: options, observe: 'response' })
      .pipe(map((res: EntityArrayResponseType) => this.convertDateArrayFromServer(res)));
  }

  delete(id: number): Observable<HttpResponse<any>> {
    return this.http.delete<any>(`${this.resourceUrl}/${id}`, { observe: 'response' });
  }

  protected convertDateFromClient(comentario: IComentario): IComentario {
    const copy: IComentario = Object.assign({}, comentario, {
      fechaCreacion:
        comentario.fechaCreacion != null && comentario.fechaCreacion.isValid() ? comentario.fechaCreacion.format(DATE_FORMAT) : null
    });
    return copy;
  }

  protected convertDateFromServer(res: EntityResponseType): EntityResponseType {
    if (res.body) {
      res.body.fechaCreacion = res.body.fechaCreacion != null ? moment(res.body.fechaCreacion) : null;
    }
    return res;
  }

  protected convertDateArrayFromServer(res: EntityArrayResponseType): EntityArrayResponseType {
    if (res.body) {
      res.body.forEach((comentario: IComentario) => {
        comentario.fechaCreacion = comentario.fechaCreacion != null ? moment(comentario.fechaCreacion) : null;
      });
    }
    return res;
  }
}
